Move previewGen and previewClear into ui
